fixed tooltip appearance issue and decided to keep scroll behavior on mobile
